Modified navigation to add a login flow. Added a navigation folder in apps where our navigation logic lies
